ワイルドカードを修正、ls.rbをwild_dir_file.rbに分割
